Removed ‘different legal name’ content
We need to remove this sentence from the ID document upload page:

“If your legal name is different to the identification document you uploaded, you will need to submit evidence of your name change too.”

The reason – we ask them to upload their change of name evidence in the previous section (personal details).

We do it there because we need to ask them whether their name appears differently on their ID documents OR their qualifications… so it needs to sit outside either of those specific sections, as a one-time “I’ve changed my name” upload.

It’s confusing if they then come to the ID interaction and we tell them about it again.

Signed-off-by: Andrew Scrivener <andrew.scrivener@digital.homeoffice.gov.uk>