Removed filename and moved error message
The document filename is removed as part of the virus scan so we need to replace with a placeholder. Using “File upload error” that will link to a content page explaining the issue as a validation message can’t be shown.

The inline error message has been moved to sit below the link too.

Signed-off-by: Andrew Scrivener <andrew.scrivener@digital.education.gov.uk>